// srm_ui/src/app/main/Models/etendering/ViewModels/rfq-header-document-text-view-model.ts
export class RFQHeaderDocumentTextViewModel {

    id: string;
    rfqId: any;

    rFQId: string;

    outputTypeId: string;

    phraseId: string;

    reference: string;

    documentText: string;

    isVisibleToSuppliers: boolean;

    categoryName: string;

    documentTextId: string;

    outputTypeName: string;

    outputTypes: any[];

    isVisibleToSupplierSave: boolean;

    attributeCategoryId: string;

    attributeCategoryName: string;

    isCategorySave: boolean;

    attributeCategorys: any[];
}